pyproject.toml
[tool.poetry]
name = "data-quality-introduction"
version = "0.2.0"
description = ""
authors = ["Gutto <guttolaudie@gmail.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
pandera = {extras = ["io"], version = "^0.20.3"}
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
duckdb = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.4"
taskipy = "^1.13.0"
isort = "^5.13.2"


[tool.poetry.group.doc.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.30"
mkdocstrings = "^0.25.1"
mkdocstrings-python = "^1.10.5"
mkdocs-mermaid2-plugin = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
ruff chec --select I --fix .
ruff format .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"
